// frontend/src/app/rest-api/types.gen.ts
export type LicensePlateResponse = {
    id?: string;
    licensePlate?: string;
    validFrom?: string;
    validTo?: string;
};

export type LicensePlateRequest = {
    licensePlate: string;
    validFrom: string;
    validTo: string;
};

export type LicensePlateCreateResponse = {
    id?: string;
    message?: string;
};

export type MessageResponse = {
    message?: string;
};
// ...
export type GetApiLicensePlatesResponse = Array<LicensePlateResponse>;

export type PostApiLicensePlatesData = {
    requestBody: LicensePlateRequest;
};

export type PostApiLicensePlatesResponse = LicensePlateCreateResponse;

export type PutApiLicensePlatesByLicensePlateIdData = {
    /**
     * ID of the license plate to delete
     */
    licensePlateId: string;
    requestBody: LicensePlateRequest;
};

export type PutApiLicensePlatesByLicensePlateIdResponse = MessageResponse;